feat: minor changes of changing component parent
